% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcapri.R
\name{readcapmod}
\alias{readcapmod}
\title{Read capri output from a gdx file}
\usage{
readcapmod(gdxfile, symName = "dataout", names = c("RALL", "EMPTY",
  "COLS", "ROWS", "Y", "VALUE"))
}
\arguments{
\item{gdxfile}{character name of a gdx file}
}
\value{
a data frame containing data from the gdx file.
}
\description{
The function \code{readcapmod()} is a wrapper around \code{gdxrrw::rgdx.param()}
}
\examples{
\dontrun{
library(gdxrrw) # Load the interface between 'GAMS' and R
# Specify the GAMS path on windows
igdx(gamsSysDir = "E:\\\\dev\\\\CAPRI\\\\GAMS\\\\win64\\\\24.7")
# Specify the GAMS path on Linux
igdx(gamsSysDir = "/opt/gams/gams25.1_linux_x64_64_sfx/")

# Specify where gdx files are located
datapath <- "E:\\\\dev\\\\CAPRI\\\\star_2.2\\\\output\\\\results\\\\capmod"

# Read one file 
firstfile <- file.path(datapath, "res_2_0810mtr_rd_ref.gdx")
res_2_0810mtr_rd_ref <- readcaprioutput(firstfile)
head(res_2_0810mtr_rd_ref)

# Read all files in the directory and combine them in a single data frame
gdxfiles <- list.files(datapath, ".gdx", full.names = TRUE)
caprilist <- lapply(gdxfiles, readcaprioutput)
capriout <- Reduce(rbind, caprilist)
head(capriout)
}
}
